Refatoração do Código. Lógica para adicionar o getStatus na classe Escalonador e aplicação dos métodos na classe.
