+++ src/main/java/br/ufpb/dcx/aps/escalonador/MaisCurto.java
@@ -1,7 +1,6 @@
 package br.ufpb.dcx.aps.escalonador;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 public class MaisCurto extends Escalonador {
@@ -12,6 +11,7 @@ public class MaisCurto extends Escalonador {
 		this.listaRetomar = new ArrayList<>();
 		this.fila = new ArrayList<>();
 		this.quantum = 0;
+		this.tipoEscalonador = TipoEscalonador.MaisCurtoPrimeiro;
 	}
 
 	public MaisCurto(int quantum) {
@@ -21,31 +21,10 @@ public class MaisCurto extends Escalonador {
 		this.listaBloqueados = new ArrayList<Processo>();
 		this.quantum = quantum;
 		this.fila = new ArrayList<Processo>();
+		this.tipoEscalonador = TipoEscalonador.MaisCurtoPrimeiro;
 	}
 
-	public String getStatus() {
-		String status = "Escalonador " + "MaisCurtoPrimeiro" + ";Processos: {";
 
-		if (rodando != null) {
-			status += "Rodando: " + rodando.toString();
-		}
-
-		if (!fila.isEmpty() && rodando == null) {
-			status += "Fila: " + this.fila.toString();
-		} else if (!fila.isEmpty() && rodando != null) {
-			status += ", Fila: " + this.fila.toString();
-		}
-
-		if (!listaBloqueados.isEmpty() && processoBloqueado == true && rodando != null) {
-			status += ", Bloqueados: " + this.listaBloqueados.toString();
-		}
-
-		if (!listaBloqueados.isEmpty() && processoBloqueado == true && rodando == null) {
-			status += "Bloqueados: " + this.listaBloqueados.toString();
-		}
-
-		return status += "};Quantum: " + this.quantum + ";Tick: " + tick;
-	}
 
 	public void tick() {
         tick++;
@@ -62,28 +41,6 @@ public class MaisCurto extends Escalonador {
             adicionarProcessoRodando();// chama o metodo para adicionar um processo a lista de rodando
         }
     }
-
-	public void adicionarProcesso(String nomeProcesso) {
-
-		if (existsProcessoByName(nomeProcesso)) {
-			throw new EscalonadorException();
-		}
-
-		if (nomeProcesso == null) {
-			throw new EscalonadorException();
-		}
-
-	}
-
-	public void adicionarProcessoRodando() {
-		if (!fila.isEmpty()) {// Só entra se a fila não estiver vazia
-			for (Processo f : fila) {// Varre a lista de fila
-				rodando = f;// adiciona o nome do processo na variavel
-				fila.remove(f);
-				break;
-			}
-		}
-	}
 
 	public void trocaRodandoParaFila() {
 		boolean bloqueado = false;
@@ -110,24 +67,6 @@ public class MaisCurto extends Escalonador {
 		}
 	}
 
-	public void adicionarProcesso(String nomeProcesso, int tempo) {
-
-		if (existsProcessoByName(nomeProcesso)) {
-			throw new EscalonadorException();
-		}
-
-	}
-
-	public void finalizarProcesso(String nomeProcesso) {
-
-		if (!existsProcessoByName(nomeProcesso) && !isRodando(nomeProcesso)) {
-			throw new EscalonadorException();
-		} else {
-			this.finalizado = true;
-			this.nomeProcessoFinalizado = nomeProcesso;
-		}
-	}
-
 	public void finalizandoProcesso(String nomeProcesso) {
 		if (rodando != null) {
 			if (rodando.getNome().equals(nomeProcesso)) {
@@ -142,39 +81,6 @@ public class MaisCurto extends Escalonador {
 			}
 		}
 	}
-
-	public void bloquearProcesso(String nomeProcesso) {
-		if (!existsProcessoByName(nomeProcesso) || !isRodando(nomeProcesso)) {
-			throw new EscalonadorException();
-		}
-		this.processoBloqueado = true;
-		nomeProcessoBloquado = nomeProcesso;
-	}
-
-	public void bloqueandoProcesso(String nomeProcessoBloq) {
-		this.processoBloqueado = false;
-		if (rodando != null) {
-			if (rodando.getNome().equals(nomeProcessoBloq)) {
-				listaBloqueados.add(rodando);
-				this.nomeProcessoBloquado = rodando.getNome();
-			}
-		}
-	}
-
-	public void retomarProcesso(String nomeProcesso) {
-
-		if (!isBloqueado(nomeProcesso) || !existsProcessoByName(nomeProcesso)) {
-			throw new EscalonadorException();
-		}
-
-		this.retomar = true;
-		for (Processo p : listaBloqueados) {
-			if (p.getNome().equals(nomeProcesso)) {
-				listaRetomar.add(p);
-			}
-		}
-	}
-
 	public void retomandoProcesso(List<Processo> retomar) {
 
 		while (!retomar.isEmpty()) {
@@ -189,59 +95,5 @@ public class MaisCurto extends Escalonador {
 			}
 		}
 	}
-
-	public Boolean existsProcessoByName(String nome) {
-		if (!fila.isEmpty()) {
-			for (Processo p : fila) {
-				if (p.getNome().equals(nome)) {
-					return true;
-				}
-			}
-		}
-		if (!listaBloqueados.isEmpty()) {
-			for (Processo b : listaBloqueados) {
-				if (b.getNome().equals(nome)) {
-					return true;
-				}
-			}
-		}
-		if (rodando != null) {
-			if (rodando.getNome() == nome) {
-				return true;
-			}
-		}
-
-		return false;
-	}
-
-	public boolean isBloqueado(String p) {
-		if (!listaBloqueados.isEmpty()) {
-			for (Processo q : listaBloqueados) {
-				if (q.getNome().equals(p)) {
-					return true;
-				}
-			}
-		}
-		return false;
-	}
-
-	public boolean isRodando(String p) {
-		if (rodando != null) {
-			if (rodando.getNome().equals(p)) {
-				return true;
-			}
-		}
-		return false;
-	}
-
-	public void adicionarProcessoTempoFixo(String nomeProcesso, int duracao) {
-        if(existsProcessoByName(nomeProcesso) || nomeProcesso == null || duracao <= 0){
-            throw new EscalonadorException();
-        }else{
-            Processo processo = new Processo(duracao, this.tick, nomeProcesso);
-            fila.add(processo);
-            Collections.sort(fila);
-        }
-    }
 }
