+++ src/main/java/br/ufpb/dcx/aps/escalonador/Prioridade.java
@@ -11,6 +11,8 @@ public class Prioridade extends Escalonador{
 			this.listaRetomar = new ArrayList<>();
 			this.fila = new ArrayList<>();
 			this.quantum = 3;
+			this.tipoEscalonador = TipoEscalonador.Prioridade;
+			
 	}
 
 	public Prioridade(int quantum) {
@@ -20,32 +22,10 @@ public class Prioridade extends Escalonador{
 		this.listaBloqueados = new ArrayList<Processo>();
 		this.quantum = quantum;
 		this.fila = new ArrayList<Processo>();
+		this.tipoEscalonador = TipoEscalonador.Prioridade;
 	}
 
-	public String getStatus() {
-		String status = "Escalonador " + "Prioridade" + ";Processos: {";
-
-		if (rodando != null) {
-			status += "Rodando: " + rodando.toString();
-		}
-
-		if (!fila.isEmpty() && rodando == null) {
-			status += "Fila: " + this.fila.toString();
-		} else if (!fila.isEmpty() && rodando != null) {
-			status += ", Fila: " + this.fila.toString();
-		}
-
-		if (!listaBloqueados.isEmpty() && processoBloqueado == true && rodando != null) {
-			status += ", Bloqueados: " + this.listaBloqueados.toString();
-		}
-
-		if (!listaBloqueados.isEmpty() && processoBloqueado == true && rodando == null) {
-			status += "Bloqueados: " + this.listaBloqueados.toString();
-		}
-
-		return status += "};Quantum: " + this.quantum + ";Tick: " + tick;
-	}
-
+	
 	public void tick() {
 		tick++;
 
@@ -92,33 +72,7 @@ public class Prioridade extends Escalonador{
 
 	}
 
-	public void adicionarProcesso(String nomeProcesso) {
-
-		if (existsProcessoByName(nomeProcesso)) {
-			throw new EscalonadorException();
-		}
-
-		if (nomeProcesso == null) {
-			throw new EscalonadorException();
-		}
-
-		if (this.quantum >= 0) {
-			Processo processo = new Processo(nomeProcesso, tick);// Cria um processo
-			tickProximoFila = quantum + processo.getTickInicial();
-			fila.add(processo);// Adiciona o processo na fila
-
-		}
-	}
-
-	public void adicionarProcessoRodando() {
-		if (!fila.isEmpty()) {// Só entra se a fila não estiver vazia
-			for (Processo f : fila) {// Varre a lista de fila
-				rodando = f;// adiciona o nome do processo na variavel
-				fila.remove(f);
-				break;
-			}
-		}
-	}
+	
 
 	public void trocaRodandoParaFila() {
 		boolean bloqueado = false;
@@ -145,34 +99,7 @@ public class Prioridade extends Escalonador{
 		}
 	}
 
-	public void adicionarProcesso(String nomeProcesso, int prioridade) {
-
-		if (existsProcessoByName(nomeProcesso)) {
-			throw new EscalonadorException();
-		}
-
-		if (nomeProcesso == null) {
-			throw new EscalonadorException();
-		}
-
-		if (this.quantum >= 0) {
-			Processo processo = new Processo(nomeProcesso, tick, prioridade);// Cria um processo
-			tickProximoFila = quantum + processo.getTickInicial();
-			fila.add(processo);// Adiciona o processo na fila
-			Collections.sort(fila);
-		}
-	}
-
-	public void finalizarProcesso(String nomeProcesso) {
-
-		if (!existsProcessoByName(nomeProcesso) && !isRodando(nomeProcesso)) {
-			throw new EscalonadorException();
-		} else {
-			this.finalizado = true;
-			this.nomeProcessoFinalizado = nomeProcesso;
-		}
-	}
-
+	
 	public void finalizandoProcesso(String nomeProcesso) {
 		if (rodando != null) {
 			if (rodando.getNome().equals(nomeProcesso)) {
@@ -188,14 +115,6 @@ public class Prioridade extends Escalonador{
 		}
 	}
 
-	public void bloquearProcesso(String nomeProcesso) {
-		if (!existsProcessoByName(nomeProcesso) || !isRodando(nomeProcesso)) {
-			throw new EscalonadorException();
-		}
-		this.processoBloqueado = true;
-		nomeProcessoBloquado = nomeProcesso;
-	}
-
 	public void bloqueandoProcesso(String nomeProcessoBloq) {
 		this.processoBloqueado = false;
 		if (rodando != null) {
@@ -206,20 +125,7 @@ public class Prioridade extends Escalonador{
 		}
 	}
 
-	public void retomarProcesso(String nomeProcesso) {
-
-		if (!isBloqueado(nomeProcesso) || !existsProcessoByName(nomeProcesso)) {
-			throw new EscalonadorException();
-		}
-
-		this.retomar = true;
-		for (Processo p : listaBloqueados) {
-			if (p.getNome().equals(nomeProcesso)) {
-				listaRetomar.add(p);
-			}
-		}
-	}
-
+	
 	public void retomandoProcesso(List<Processo> retomar) {
 
 		while (!retomar.isEmpty()) {
@@ -233,53 +139,6 @@ public class Prioridade extends Escalonador{
 				}
 			}
 		}
-	}
-
-	public Boolean existsProcessoByName(String nome) {
-		if (!fila.isEmpty()) {
-			for (Processo p : fila) {
-				if (p.getNome().equals(nome)) {
-					return true;
-				}
-			}
-		}
-		if (!listaBloqueados.isEmpty()) {
-			for (Processo b : listaBloqueados) {
-				if (b.getNome().equals(nome)) {
-					return true;
-				}
-			}
-		}
-		if (rodando != null) {
-			if (rodando.getNome() == nome) {
-				return true;
-			}
-		}
-
-		return false;
-	}
-
-	public boolean isBloqueado(String p) {
-		if (!listaBloqueados.isEmpty()) {
-			for (Processo q : listaBloqueados) {
-				if (q.getNome().equals(p)) {
-					return true;
-				}
-			}
-		}
-		return false;
-	}
-
-	public boolean isRodando(String p) {
-		if (rodando != null) {
-			if (rodando.getNome().equals(p)) {
-				return true;
-			}
-		}
-		return false;
-	}
-
-	public void adicionarProcessoTempoFixo(String string, int duracao) {
 	}
 
 }
