+++ src/main/java/br/ufpb/dcx/aps/escalonador/RoundRobin.java
@@ -1,7 +1,6 @@
 package br.ufpb.dcx.aps.escalonador;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 public class RoundRobin extends Escalonador{
@@ -11,6 +10,7 @@ public class RoundRobin extends Escalonador{
             this.listaRetomar = new ArrayList<>();
             this.fila = new ArrayList<>();
             this.quantum = 3;
+            this.tipoEscalonador = TipoEscalonador.RoundRobin;
     }
 
     public RoundRobin(int quantum) {
@@ -20,31 +20,10 @@ public class RoundRobin extends Escalonador{
         this.listaBloqueados = new ArrayList<Processo>();
         this.quantum = quantum;
         this.fila = new ArrayList<Processo>();
+        this.tipoEscalonador = TipoEscalonador.RoundRobin;
     }
 
-    public String getStatus() {
-        String status = "Escalonador " + "RoundRobin" + ";Processos: {";
-
-        if (rodando != null) {
-            status += "Rodando: " + rodando.toString();
-        }
-
-        if (!fila.isEmpty() && rodando == null) {
-            status += "Fila: " + this.fila.toString();
-        } else if (!fila.isEmpty() && rodando != null) {
-            status += ", Fila: " + this.fila.toString();
-        }
-
-        if (!listaBloqueados.isEmpty() && processoBloqueado == true && rodando != null) {
-            status += ", Bloqueados: " + this.listaBloqueados.toString();
-        }
-
-        if (!listaBloqueados.isEmpty() && processoBloqueado == true && rodando == null) {
-            status += "Bloqueados: " + this.listaBloqueados.toString();
-        }
-
-        return status += "};Quantum: " + this.quantum + ";Tick: " + tick;
-    }
+  
 
     public void tick() {
         tick++;
@@ -82,34 +61,6 @@ public class RoundRobin extends Escalonador{
         }
     }
 
-    public void adicionarProcesso(String nomeProcesso) {
-
-        if (existsProcessoByName(nomeProcesso)) {
-            throw new EscalonadorException();
-        }
-
-        if (nomeProcesso == null) {
-            throw new EscalonadorException();
-        }
-
-        if (this.quantum >= 0) {
-            Processo processo = new Processo(nomeProcesso, tick);// Cria um processo
-            tickProximoFila = quantum + processo.getTickInicial();
-            fila.add(processo);// Adiciona o processo na fila
-
-        }
-    }
-
-    public void adicionarProcessoRodando() {
-        if (!fila.isEmpty()) {// Só entra se a fila não estiver vazia
-            for (Processo f : fila) {// Varre a lista de fila
-                rodando = f;// adiciona o nome do processo na variavel
-                fila.remove(f);
-                break;
-            }
-        }
-    }
-
     public void trocaRodandoParaFila() {
         boolean bloqueado = false;
 
@@ -129,35 +80,6 @@ public class RoundRobin extends Escalonador{
         }
     }
 
-    public void adicionarProcesso(String nomeProcesso, int prioridade) {
-
-        if (existsProcessoByName(nomeProcesso)) {
-            throw new EscalonadorException();
-        }
-
-        if (nomeProcesso == null) {
-            throw new EscalonadorException();
-        }
-
-        if (this.quantum >= 0) {
-            Processo processo = new Processo(nomeProcesso, tick, prioridade);// Cria um processo
-            tickProximoFila = quantum + processo.getTickInicial();
-            fila.add(processo);// Adiciona o processo na fila
-            Collections.sort(fila);
-
-        }
-    }
-
-    public void finalizarProcesso(String nomeProcesso) {
-
-        if (!existsProcessoByName(nomeProcesso) && !isRodando(nomeProcesso)) {
-            throw new EscalonadorException();
-        } else {
-            this.finalizado = true;
-            this.nomeProcessoFinalizado = nomeProcesso;
-        }
-    }
-
     public void finalizandoProcesso(String nomeProcesso) {
         if (rodando != null) {
             if (rodando.getNome().equals(nomeProcesso)) {
@@ -173,13 +95,6 @@ public class RoundRobin extends Escalonador{
         }
     }
 
-    public void bloquearProcesso(String nomeProcesso) {
-        if (!existsProcessoByName(nomeProcesso) || !isRodando(nomeProcesso)) {
-            throw new EscalonadorException();
-        }
-        this.processoBloqueado = true;
-        nomeProcessoBloquado = nomeProcesso;
-    }
 
     public void bloqueandoProcesso(String nomeProcessoBloq) {
         this.processoBloqueado = false;
@@ -187,20 +102,6 @@ public class RoundRobin extends Escalonador{
             if (rodando.getNome().equals(nomeProcessoBloq)) {
                 listaBloqueados.add(rodando);
                 this.nomeProcessoBloquado = rodando.getNome();
-            }
-        }
-    }
-
-    public void retomarProcesso(String nomeProcesso) {
-
-        if (!isBloqueado(nomeProcesso) || !existsProcessoByName(nomeProcesso)) {
-            throw new EscalonadorException();
-        }
-
-        this.retomar = true;
-        for (Processo p : listaBloqueados) {
-            if (p.getNome().equals(nomeProcesso)) {
-                listaRetomar.add(p);
             }
         }
     }
@@ -218,52 +119,5 @@ public class RoundRobin extends Escalonador{
                 }
             }
         }
-    }
-
-    public Boolean existsProcessoByName(String nome) {
-        if (!fila.isEmpty()) {
-            for (Processo p : fila) {
-                if (p.getNome().equals(nome)) {
-                    return true;
-                }
-            }
-        }
-        if (!listaBloqueados.isEmpty()) {
-            for (Processo b : listaBloqueados) {
-                if (b.getNome().equals(nome)) {
-                    return true;
-                }
-            }
-        }
-        if (rodando != null) {
-            if (rodando.getNome() == nome) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    public boolean isBloqueado(String p) {
-        if (!listaBloqueados.isEmpty()) {
-            for (Processo q : listaBloqueados) {
-                if (q.getNome().equals(p)) {
-                    return true;
-                }
-            }
-        }
-        return false;
-    }
-
-    public boolean isRodando(String p) {
-        if (rodando != null) {
-            if (rodando.getNome().equals(p)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    public void adicionarProcessoTempoFixo(String string, int duracao) {
     }
 }
